chore: refactor file structure
