+++ packages/web/src/features/register-club/components/_atomic/RulesButton.tsx
