+++ packages/web/src/features/register-club/constants/registerClub.ts
@@ -13,3 +13,10 @@ export const DEPARTMENT_AUTONOMY_RULES = `제54조(분과자치규칙)
 ⑥ 분과학생회장은 매 정규학기 개강일 20일 이전까지 분과자치규칙 존재여부 및 최신 분과
 자치규칙을 집행부에 송부해야 한다.
 `;
+
+export const CLUBS_COMPLIANCE_LIST = [
+  "1. 동아리연합회칙을 준수하겠습니다.",
+  "2. KAIST의 대학 문화 발전을 위해 이바지하겠습니다.",
+  "3. 본 회의 민주적 의사결정에 성실히 참여하겠습니다.",
+  "4. 분과자치규칙을 준수하겠습니다.",
+];
